+++ reconcile/test/fixtures/ocm/oidc/get_full.yml
@@ -1,6 +1,6 @@
 ---
 expected_return_value:
-  cluster: cluster-1
+  cluster: test-cluster
   name: oidc-auth
   id: idp-id-2
   client_id: another-client-id
@@ -15,18 +15,18 @@ expected_return_value:
 
 urls:
   - name: list idps
-    uri: /api/clusters_mgmt/v1/clusters/cluster-1-id/identity_providers
+    uri: /api/clusters_mgmt/v1/clusters/osd-cluster-id/identity_providers
     method: get
     responses:
       - kind: IdentityProviderList
-        href: /api/clusters_mgmt/v1/clusters/cluster-1-id/identity_providers
+        href: /api/clusters_mgmt/v1/clusters/osd-cluster-id/identity_providers
         page: 1
         size: 2
         total: 2
         items:
           - kind: IdentityProvider
             type: GithubIdentityProvider
-            href: /api/clusters_mgmt/v1/clusters/cluster-1-id/identity_providers/idp-id-1
+            href: /api/clusters_mgmt/v1/clusters/osd-cluster-id/identity_providers/idp-id-1
             id: idp-id-1
             name: github-auth
             mapping_method: claim
@@ -36,7 +36,7 @@ urls:
                 - just-a-team-name
           - kind: IdentityProvider
             type: OpenIDIdentityProvider
-            href: /api/clusters_mgmt/v1/clusters/cluster-1-id/identity_providers/idp-id-2
+            href: /api/clusters_mgmt/v1/clusters/osd-cluster-id/identity_providers/idp-id-2
             id: idp-id-2
             name: oidc-auth
             mapping_method: claim
